algorithmArt-main/src/main/java/de/postlab/alg/comparisonsort/insertionsort/StdInsertionsortSmpl.java
package de.postlab.alg.comparisonsort.insertionsort;

import de.postlab.alg.util.SortDebugHelper;
import lombok.Getter;

/**
 * AlgorithmArt - A collection of algorithm implementations,
 * tests and profiling programs to research the algorithms.
 * <p>
 * (c) Copyright 2015 by Michael Krämer
 * See https://github.com/mkraemerx/algorithm-art
 * <p>
 * This file is part of AlgorithmArt.
 * <p>
 * AlgorithmArt is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 * <p>
 * Foobar is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 * <p>
 * You should have received a copy of the GNU General Public License
 * along with Foobar. If not, see <http://www.gnu.org/licenses/>.
 * <p>
 * Initially created by michel on 23.06.15.
 */
public class StdInsertionsortSmpl {

    @Getter
    private SortDebugHelper helper;

    public StdInsertionsortSmpl() {
        helper = new SortDebugHelper();
    }

    public void sort(int[] arr) {
        insertionSort(arr);
    }

    public void insertionSort(int[] arr) {
        for (int i=1; i<arr.length; i++) {
            for (int j=i; j>0; j--) {
                helper.comparisons++;
                if (arr[j] < arr[j-1]) {
                    int tmp = arr[j-1];
                    arr[j-1] = arr[j];
                    arr[j] = tmp;
                } else {
                    break;
                }
            }
        }
    }
}
